requirements.txt
flask
google
google-api-python-client
PyPDF2
python-dotenv
google-generativeai